More small improvements.
I stopped doing syntax highlighting with js. This reduced the script
count to 1 which is neat.

Everything is static now makes things easier to maintain.

I moved everything around just a little, adjusting colors and font sizes
as I saw fit.

Changed the homepage so it uses the full height of the screen